feat: add JsonValue constructors for all number types
Now that we implicitly accept floats in the JsonValue constructor
and implicit conversion, integer types now get converted to a float
rather than a double. To preserve the pre-existing functionality
for all numbers other than floats, we need constructors and
conversions for every integer type where we explicitly use the
double constructor.